Update bpm-sdk (#2)
* Create docker command as part of the configuration

This allows overwriting the command for troubleshooting or other
tinkering purposes. It also simplifies the plugin code that starts the
container.

* Add in/out peers configuration option

* Update bpm-sdk

* Renamed things to work with changes in the SDK

* Fix default baseDir

* Add status command

Note: It is currently a bit cumbersone to write all of these commands
for each plugin. We will work on simplifying this by providing more
default and help functions.

* Re-enabling vendor temporarily

This is just until we've fully assessed the security implications

* Remove leftover from merge

* Linter